Full_Project_OCR/requirement.txt
mmcv==2.0.1
mmdet==3.1.0
mmengine==0.10.3
Levenshtein==0.23.0
pandas==1.5.3
numpy==1.23.5
jiwer==3.0.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
torch==2.1.2
pyspellchecker==0.8.1
indexer